Possible reusable code
I'm aware Teesside University have recently done research on IoT systems for hospital equipment tracking. Particularly in database architecture.

I've took the liberty of emailing the university to see if they have any code or research that could be deemed relevant to this project.

Apologies for using this channel, I was unable to find the slack channel.